Refactor the SchemaModel class with a base class
