+++ frontend/packages/schema-model/src/lib/SchemaModel/index.ts
@@ -1,0 +1,1 @@
+export { SchemaModel } from './SchemaModel';
